Set editor and add .vim
I set the editor to be Vim in .bash_profile so that `bundle open [gem]`
works. I moved .bash_profile and .vim into my dotfiles directory and
made a symlink so that they can also be found in the root directory by
running `ln -s [file_path] .` in the root directory.